+++ json2dart_db/android/app/src/main/kotlin/com/awesome/json2dart_db/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.awesome.json2dart_db
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}