+++ json2dart_db/lib/database/base_db_manager.dart
@@ -3,6 +3,9 @@ import 'dart:async';
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
 
+import 'model/column_info.dart';
+import 'model/table_info.dart';
+
 abstract class BaseDbManager {
   static BaseDbManager? _instance;
   Database? _database;
@@ -28,6 +31,7 @@ abstract class BaseDbManager {
     await onCreate(db, newVersion);
   }
 
+  /// onCreate the database
   FutureOr<void> onCreate(Database db, int version) async {}
 
   ///项目进行初始化
@@ -48,4 +52,54 @@ abstract class BaseDbManager {
 
   ///数据库名称，可复写
   String getDbName() => 'client_db.db';
+
+  ///查询数据库中所有的表
+  Future<List<String>> queryTables() async {
+    List<Map<String, Object?>> maps =
+        await _database!.rawQuery("select name from sqlite_master where type='table' order by name;");
+    List<String> _tables = [];
+    for (Map<String, Object?> map in maps) {
+      _tables.add(map['name'].toString());
+    }
+    return _tables;
+  }
+
+  ///查询表信息
+  Future<TableInfo?> queryTableInfo(String name) async {
+    List<Map<String, Object?>> maps = await _database!.rawQuery('select * from sqlite_master where name = "$name"');
+    if (maps.isNotEmpty) {
+      TableInfo tableInfo = TableInfo.fromJson(maps[0]);
+      List<Map<String, Object?>> datas = await _database!.rawQuery("pragma table_info ('$name');");
+      List<ColumnInfo> columns = [];
+      for (Map<String, Object?> map in datas) {
+        ColumnInfo info = ColumnInfo.fromJson(map);
+        columns.add(info);
+      }
+      tableInfo.columns = columns;
+      return tableInfo;
+    }
+    return null;
+  }
+
+  ///删除数据表的功能
+  Future<void> drop(String tableName) async {
+    return _database!.execute("DROP TABLE $tableName");
+  }
+
+  ///查询该表的数据数
+  Future<int> queryCount(String tableName) async {
+    List<Map<String, Object?>> data = await _database!.rawQuery('SELECT count(*) FROM $tableName');
+    return data[0]['count(*)'] as int;
+  }
+
+  ///删除所有数据
+  Future<void> clearTable(String tableName) async {
+    try {
+      await _database!.execute("delete from $tableName");
+      //reset the auto increase id
+      await _database!.execute("update sqlite_sequence SET seq = 0 where name ='$tableName';");
+    } catch (e) {
+      print(e);
+    }
+  }
 }