+++ json2dart_db/CHANGELOG.md
@@ -1,5 +1,12 @@
+## 0.0.5
+
+- 1.support query database tables
+- 2.support query tables columns
+
 ## 0.0.2
+
 - Support to upgrade the tables
+
 ## 0.0.1
 
 - Initial version,use json2dart_safe together,